requirements.txt
pyTelegramBotAPI
google-generativeai
pillow